+++ quickwit/quickwit-indexing/src/actors/cooperative_indexing.rs
