+++ quickwit/quickwit-indexing/src/actors/index_serializer.rs
