+++ quickwit/quickwit-indexing/src/actors/merge_executor.rs
