+++ quickwit/quickwit-indexing/src/actors/merge_scheduler_service.rs
