+++ quickwit/quickwit-indexing/src/actors/packager.rs
