+++ quickwit/quickwit-indexing/src/lib.rs
@@ -26,9 +26,8 @@ use tracing::info;
 
 use crate::actors::MergeSchedulerService;
 pub use crate::actors::{
-    FinishPendingMergesAndShutdownPipeline, IndexingError, IndexingPipeline,
-    IndexingPipelineParams, IndexingService, BoxPipelineHandle, Sequencer,
-    SplitsUpdateMailbox,
+    BoxPipelineHandle, FinishPendingMergesAndShutdownPipeline, IndexingError, IndexingPipeline,
+    IndexingPipelineParams, IndexingService, Sequencer, SplitsUpdateMailbox,
 };
 pub use crate::controlled_directory::ControlledDirectory;
 use crate::models::IndexingStatistics;
